pyproject.toml
[project]
name = "contrail-segmentation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Aryan Garg", email = "aryangarg794@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "dill>=0.4.1",
    "goes2go>=2025.10.0",
    "ipykernel>=7.2.0",
    "lightning>=2.6.1",
    "matplotlib>=3.10.8",
    "pyyaml>=6.0.3",
    "scikit-learn>=1.8.0",
    "segmentation-models-pytorch>=0.5.0",
    "timm>=1.0.25",
    "torch>=2.10.0",
    "torchvision>=0.25.0",
    "tqdm>=4.67.3",
    "typing>=3.10.0.0",
    "wandb>=0.25.0",
]

[project.scripts]
contrail-segmentation = "contrail_segmentation:main"

[build-system]
requires = ["uv_build>=0.9.3,<0.10.0"]
build-backend = "uv_build"

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[tool.uv.sources]
torch = [
  { index = "pytorch-cu128", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
torchvision = [
  { index = "pytorch-cu128", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
